+++ CHANGELOG.md
@@ -23,7 +23,8 @@ Security in case of vulnerabilities.
 - Ability to edit project information for conversions and transfers
 - ProjectGroup is created when a project is edited with a new GRN
 - Added `Add, Edit and Delete External Contacts for both conversion and transfer projects.
-- Added `Confirm this transfer has authority to proceed` task
+- Added `Confirm this transfer has authority to proceed` task.
+- Added `Confirm the date the academy opened` task.
 
 ### Changed
 - Group reference number links to the group on "About the project"